Gettting data figured out
Had to figure out how i would structure the data and how the sessions
would work. Had a heck of a time getting the user methods for sequelize
through my thick head. Finally have that going and data going back and
forth and into the view after also wrestling with ejs a bit.